+++ client/src/utility/templateObjects/configTemplate.ts
@@ -28,4 +28,5 @@ export const configTemplate: Config = {
   monthSchema:
     'January;February;March;April;May;June;July;August;September;October;November;December',
   showTime: false,
+  defaultTheme: 'tron',
 };